jwt 필터 및 인증 구현
비밀키 길이 문제 해결 요망.

+++ ASSA_backend/assa-backend/src/main/kotlin/com/assa/assabackend/entity/AppUser.kt
@@ -18,7 +18,7 @@ data class AppUser(
     val regionMetroId: Long? = null,
 
     @Column(name = "region_district_id")
-    val regionDistrictId: Int? = null,
+    val regionDistrictId: Long? = null,
 
     @Column(name = "name", length = 30, nullable = false)
     val name: String,
